# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "猫影短剧",
    "version": "1.0.0"
  },
  "paths": {
    "/api/novel": {
      "post": {
        "tags": [
          "小说/剧本管理"
        ],
        "summary": "创建小说/剧本",
        "operationId": "create_novel_api_novel_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovelCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_NovelOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "小说/剧本管理"
        ],
        "summary": "获取小说/剧本列表",
        "operationId": "get_novel_list_api_novel_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "页码：默认为1，必须为≥1的整数。示例：?page=2",
              "default": 1,
              "title": "Page"
            },
            "description": "页码：默认为1，必须为≥1的整数。示例：?page=2"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "每页数量：默认为10，范围1-100。示例：?page_size=20",
              "default": 10,
              "title": "Page Size"
            },
            "description": "每页数量：默认为10，范围1-100。示例：?page_size=20"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\n        排序规则：格式为「字段名+符号」（无符号=升序，-=降序），多字段用逗号分隔。\n        示例：?sort=create_time,-id → 按create_time升序、id降序排列\n        ",
              "title": "Sort"
            },
            "description": "\n        排序规则：格式为「字段名+符号」（无符号=升序，-=降序），多字段用逗号分隔。\n        示例：?sort=create_time,-id → 按create_time升序、id降序排列\n        "
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\n        1. 搜索：直接传关键词（会对预设的可搜索字段进行模糊匹配）\n           示例：?search=张三 → 搜索所有含\"张三\"的记录\n           \n        2. 过滤：按「字段__操作符=值」格式传参，支持多条件（用&连接）\n           常用操作符：\n           - __contains: 包含（字符串模糊匹配）\n           - __lte: 小于等于（数字/日期）\n           - __gte: 大于等于（数字/日期）\n           - __gt: 大于（数字/日期）\n           - __lt: 小于（数字/日期）\n           \n           示例：\n           ?search=张三&age__gte=18 → 姓名含\"张三\"且年龄≥18\n           ?create_time__gte=2023-01-01 → 创建时间在2023-01-01之后\n        ",
              "title": "Search"
            },
            "description": "\n        1. 搜索：直接传关键词（会对预设的可搜索字段进行模糊匹配）\n           示例：?search=张三 → 搜索所有含\"张三\"的记录\n           \n        2. 过滤：按「字段__操作符=值」格式传参，支持多条件（用&连接）\n           常用操作符：\n           - __contains: 包含（字符串模糊匹配）\n           - __lte: 小于等于（数字/日期）\n           - __gte: 大于等于（数字/日期）\n           - __gt: 大于（数字/日期）\n           - __lt: 小于（数字/日期）\n           \n           示例：\n           ?search=张三&age__gte=18 → 姓名含\"张三\"且年龄≥18\n           ?create_time__gte=2023-01-01 → 创建时间在2023-01-01之后\n        "
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_PaginationResponse_NovelBriefOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novel/{novel_id}": {
      "put": {
        "tags": [
          "小说/剧本管理"
        ],
        "summary": "全量修改小说/剧本",
        "operationId": "update_novel_api_novel__novel_id__put",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovelUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_NovelOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "小说/剧本管理"
        ],
        "summary": "局部更新小说/剧本",
        "operationId": "patch_novel_api_novel__novel_id__patch",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovelPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_NovelOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "小说/剧本管理"
        ],
        "summary": "获取小说/剧本详情",
        "operationId": "get_novel_api_novel__novel_id__get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_NovelOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "小说/剧本管理"
        ],
        "summary": "删除一个小说/剧本",
        "operationId": "delete_novel_api_novel__novel_id__delete",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novel/{novel_id}/split": {
      "get": {
        "tags": [
          "小说/剧本管理"
        ],
        "summary": "使用nlp智能拆分章节",
        "operationId": "split_novel_api_novel__novel_id__split_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_NovelOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/video/generate/": {
      "post": {
        "tags": [
          "视频生成"
        ],
        "summary": "提交视频生成",
        "description": "提交视频生成请求，返回 Video 记录",
        "operationId": "generate_video_api_video_generate__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_VideoOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/video/query/{video_id}": {
      "get": {
        "tags": [
          "视频生成"
        ],
        "summary": "查询视频生成状态",
        "description": "轮询查询视频生成进度",
        "operationId": "query_video_api_video_query__video_id__get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_VideoQueryOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/video": {
      "get": {
        "tags": [
          "视频生成"
        ],
        "summary": "获取视频列表",
        "operationId": "get_video_list_api_video_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "页码：默认为1，必须为≥1的整数。示例：?page=2",
              "default": 1,
              "title": "Page"
            },
            "description": "页码：默认为1，必须为≥1的整数。示例：?page=2"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "每页数量：默认为10，范围1-100。示例：?page_size=20",
              "default": 10,
              "title": "Page Size"
            },
            "description": "每页数量：默认为10，范围1-100。示例：?page_size=20"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\n        排序规则：格式为「字段名+符号」（无符号=升序，-=降序），多字段用逗号分隔。\n        示例：?sort=create_time,-id → 按create_time升序、id降序排列\n        ",
              "title": "Sort"
            },
            "description": "\n        排序规则：格式为「字段名+符号」（无符号=升序，-=降序），多字段用逗号分隔。\n        示例：?sort=create_time,-id → 按create_time升序、id降序排列\n        "
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\n        1. 搜索：直接传关键词（会对预设的可搜索字段进行模糊匹配）\n           示例：?search=张三 → 搜索所有含\"张三\"的记录\n           \n        2. 过滤：按「字段__操作符=值」格式传参，支持多条件（用&连接）\n           常用操作符：\n           - __contains: 包含（字符串模糊匹配）\n           - __lte: 小于等于（数字/日期）\n           - __gte: 大于等于（数字/日期）\n           - __gt: 大于（数字/日期）\n           - __lt: 小于（数字/日期）\n           \n           示例：\n           ?search=张三&age__gte=18 → 姓名含\"张三\"且年龄≥18\n           ?create_time__gte=2023-01-01 → 创建时间在2023-01-01之后\n        ",
              "title": "Search"
            },
            "description": "\n        1. 搜索：直接传关键词（会对预设的可搜索字段进行模糊匹配）\n           示例：?search=张三 → 搜索所有含\"张三\"的记录\n           \n        2. 过滤：按「字段__操作符=值」格式传参，支持多条件（用&连接）\n           常用操作符：\n           - __contains: 包含（字符串模糊匹配）\n           - __lte: 小于等于（数字/日期）\n           - __gte: 大于等于（数字/日期）\n           - __gt: 大于（数字/日期）\n           - __lt: 小于（数字/日期）\n           \n           示例：\n           ?search=张三&age__gte=18 → 姓名含\"张三\"且年龄≥18\n           ?create_time__gte=2023-01-01 → 创建时间在2023-01-01之后\n        "
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_PaginationResponse_VideoBriefOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/video/{video_id}": {
      "get": {
        "tags": [
          "视频生成"
        ],
        "summary": "获取视频详情",
        "operationId": "get_video_api_video__video_id__get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema_VideoOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "视频生成"
        ],
        "summary": "删除视频",
        "operationId": "delete_video_api_video__video_id__delete",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NovelCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name",
            "description": "小说名称"
          },
          "author": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Author",
            "description": "作者"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "描述"
          },
          "cover": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover",
            "description": "封面图URL"
          },
          "total_chapters": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Chapters",
            "description": "总章节数"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content",
            "description": "正文内容"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "NovelCreate",
        "description": "创建请求：name 必填"
      },
...
      "NovelPatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "小说名称"
          },
          "author": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Author",
            "description": "作者"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "描述"
          },
          "cover": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover",
            "description": "封面图URL"
          },
          "total_chapters": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Chapters",
            "description": "总章节数"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content",
            "description": "正文内容"
          }
        },
        "type": "object",
        "title": "NovelPatch",
        "description": "局部更新：全字段可选"
      },
      "NovelUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name",
            "description": "小说名称"
          },
          "author": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Author",
            "description": "作者"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "描述"
          },
          "cover": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover",
            "description": "封面图URL"
          },
          "total_chapters": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Chapters",
            "description": "总章节数"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content",
            "description": "正文内容"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "NovelUpdate",
        "description": "全量更新：逻辑同创建"
      },
...
      "ResponseSchema": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "data": {
            "anyOf": [
              {

              },
              {
                "items": {

                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "操作成功"
          }
        },
        "type": "object",
        "title": "ResponseSchema",
        "description": "统一API响应格式\n\n属性:\n    code (int): 响应状态码，0表示成功，非0表示异常\n    data (T): 响应数据，泛型类型\n    message (str): 响应消息",
        "example": {
          "code": 0,
          "data": {

          },
          "message": "操作成功"
        }
      },
...
      "ResponseSchema_NovelOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NovelOut"
              },
              {
                "items": {

                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "操作成功"
          }
        },
        "type": "object",
        "title": "ResponseSchema[NovelOut]",
        "example": {
          "code": 0,
          "data": {

          },
          "message": "操作成功"
        }
      },
...
      "ResponseSchema_PaginationResponse_NovelBriefOut__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationResponse_NovelBriefOut_"
              },
              {
                "items": {

                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "操作成功"
          }
        },
        "type": "object",
        "title": "ResponseSchema[PaginationResponse[NovelBriefOut]]",
        "example": {
          "code": 0,
          "data": {

          },
          "message": "操作成功"
        }
      },
...
      "ResponseSchema_PaginationResponse_VideoBriefOut__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationResponse_VideoBriefOut_"
              },
              {
                "items": {

                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "操作成功"
          }
        },
        "type": "object",
        "title": "ResponseSchema[PaginationResponse[VideoBriefOut]]",
        "example": {
          "code": 0,
          "data": {

          },
          "message": "操作成功"
        }
      },
...
      "ResponseSchema_VideoOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VideoOut"
              },
              {
                "items": {

                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "操作成功"
          }
        },
        "type": "object",
        "title": "ResponseSchema[VideoOut]",
        "example": {
          "code": 0,
          "data": {

          },
          "message": "操作成功"
        }
      },
      "ResponseSchema_VideoQueryOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VideoQueryOut"
              },
              {
                "items": {

                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "操作成功"
          }
        },
        "type": "object",
        "title": "ResponseSchema[VideoQueryOut]",
        "example": {
          "code": 0,
          "data": {

          },
          "message": "操作成功"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VideoGenerateRequest": {
        "properties": {
          "scene_id": {
            "type": "integer",
            "title": "Scene Id",
            "description": "分镜ID"
          },
          "model_type": {
            "$ref": "#/components/schemas/VideoModelTypeEnum"
          }
        },
        "type": "object",
        "required": [
          "scene_id",
          "model_type"
        ],
        "title": "VideoGenerateRequest",
        "description": "提交视频生成请求"
      },
...
    }
  }
}